consolidate http request handling, and use api urls fresh every time (#30)
api urls were loaded into each manager's httpclient in the constructors. this resulted in tweaks to the urls not actually changing which url was used, until the plugin is reloaded. this also introduced a problem where if the url is invalid, then the plugin will fail to load outright, and the user will have to manually edit save data to fix the url. so obviously both of those things suck XD